+++ CHANGELOG.md
@@ -59,6 +59,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   or array of string patterns, or with the repeatable `--remote` flag,
   which also accepts string patterns.
 
+* The default target revisions for `jj git push` can now be configured via
+  `revsets.git-push`.
+
 * `jj run` now uses the sparse patterns from the workspace it's run from.
   Use the `--sparse-patterns` option to control this behavior (evaluated
   per each `jj run` invocation).